requirements.txt
openai>=1.12.0
langchain>=0.1.6
langchain-openai>=0.0.5
chromadb>=0.4.22
pypdf2>=3.0.1
python-dotenv>=1.0.0
tiktoken>=0.5.2
sentence-transformers>=2.3.1
numpy>=1.24.0
pydantic>=2.6.0
google-generativeai
httpx>=0.25.0
fastapi
uvicorn[standard]
python-multipart
